# Improve Vita performance to ~60 fps