--- a/src/APU.h
+++ b/src/APU.h
@@ -19,18 +19,11 @@ public:
 	void triggerChannel3();
 	void triggerChannel4();
 
-	void updateChannel1Timer(uint8_t data);
-	void updateChannel2Timer(uint8_t data);
-	void updateChannel3Timer(uint8_t data);
-	void updateChannel4Timer(uint8_t data);
-
 private:
 	void updateChannel1();
 	void updateChannel2();
 	void updateChannel3();
 	void updateChannel4();
-
-	void updateControl();
 
 	void updateFrameSequencer();
 
@@ -49,21 +42,15 @@ private:
 	float volume = 0.001f;
 	uint8_t soundOn = 1;
 
-	// Left and right sound output data
-	uint8_t volumeSO2 = 0;
-	uint8_t volumeSO1 = 0;
-	uint8_t selectionSO2[4] = {};
-	uint8_t selectionSO1[4] = {};
-
 	uint16_t readPos = 0;
 	uint16_t writePos = 0;
-	
+
 	// TODO: Consider switching channel operation to be more object oriented
 	struct channel {
 		float buffer[1024] = {};
 		uint8_t soundOn = 0;
 		uint8_t volume = 0;
-		float sample = 0.0f;
+		uint8_t sample = 0;		// (0 to 16 for -1.0 to 1.0)
 
 		uint32_t lengthCounter = 0;
 		uint32_t sweepCounter = 0;
@@ -71,12 +58,17 @@ private:
 		uint32_t frequencyCounter = 0;
 		uint16_t bufferIndex = 0;
 		uint8_t waveIndex = 0;
-	};
 
-	channel channel1;
-	channel channel2;
-	channel channel3;
-	channel channel4;
+		uint8_t shiftAmount = 0;
+		uint8_t divisor = 0;
+		uint8_t widthMode = 0;
+
+		uint32_t frequencyPeriod = 0;
+
+		uint8_t dacPower = 0;
+
+		uint8_t bufferSample = 0;
+	};
 
 	uint16_t shiftRegister = 0xFFFF;
 
@@ -90,14 +82,21 @@ private:
 	// Various arrays to be indexed
 	uint8_t squareWaveRatio[4] = { 1, 2, 4, 6 };
 	uint8_t waveVolume[4] = { 0, 0x0F, 0x07, 0x03 };		// 0%, 100%, 50%, 25%
-	uint8_t noiseDivisor[8] = { 8, 16, 32, 48, 64, 80, 96, 112 };
+	//uint8_t noiseDivisor[8] = { 8, 16, 32, 48, 64, 80, 96, 112 };
 
 public:
-	uint8_t dacPower1 = 0;
-	uint8_t dacPower2 = 0;
-	uint8_t dacPower3 = 0;
-	uint8_t dacPower4 = 0;
+	channel channel1;
+	channel channel2;
+	channel channel3;
+	channel channel4;
 
+	// Left and right sound output data
+	uint8_t volumeSO2 = 0;
+	uint8_t volumeSO1 = 0;
+	uint8_t selectionSO2[4] = {};
+	uint8_t selectionSO1[4] = {};
+
+	uint8_t noiseDivisor[8] = { 8, 16, 32, 48, 64, 80, 96, 112 };
 
 public:
 	// Debug related information