I finally managed to increase vita performance to around 60 fps in Super Mario Land 1-1. At times it can dip to around ~57 fps, I think when there is more audio to execute.

The performance improvements I've made are as follows:
- temp

For the future, more performance improvements can be made to maintain a more stable 60 fps. As well, many games may not run at 60 fps, I've only tested Super Mario Land 1-1.